dev enum rule comparator

+++ jedani-worker/src/main/java/ru/complitex/domain/entity/Status.java
@@ -7,16 +7,16 @@ import ru.complitex.common.entity.IdEnum;
  * 30.11.2017 15:31
  */
 public enum Status implements IdEnum {
-    INACTIVE(0), ACTIVE(1),  ARCHIVE(2), SYNC(3), SYSTEM(4);
+    INACTIVE(0L), ACTIVE(1L),  ARCHIVE(2L), SYNC(3L), SYSTEM(4L);
 
-    private Integer id;
+    private Long id;
 
-    Status(Integer id) {
+    Status(Long id) {
         this.id = id;
     }
 
     @Override
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 }
